refactor: move utilities to utils package
