+++ source/modulo_components/modulo_components/utilities/__init__.py
@@ -1,1 +1,1 @@
-from modulo_components.utilities.utilities import parse_topic_name, modify_parameter_overrides
+from modulo_components.utilities.utilities import modify_parameter_overrides
