+++ source/modulo_utils/include/modulo_utils/predicate_variant.hpp
@@ -2,8 +2,8 @@
 
 #include <variant>
 
-namespace modulo_components::utilities {
+namespace modulo_utils {
 
 typedef std::variant<bool, std::function<bool(void)>> PredicateVariant;
 
-}// namespace modulo_components::utilities
+}// namespace modulo_components::modulo_utils
